Changed label sizes, added basic click functionality
Need to figure out a way to attach each boolean value to the individual
numbers. Also need to convert my declarations of components to a for
loop, but that's obvious and will be done later (if anyone reviews the
git log, I promise I'm not a dummy programmer - its just easier to
visualize by declaring everything alone to begin with). Also need to
clean up and tidy.